Telephony: Add an API for getting recipient address of SubmitPdu
There is no API to get the recipient address of submit pdu for SIM
card message. getDisplayOriginatingAddress() return null for submit
pdu, it will cause applicatoin can not get address of sent message.
Thread ID can not be created for this invalid message record.

Add an API getRecipientAddress() for SubmitPdu in SmsMessage.

Change-Id: I8b8af6002b5b309cba5383adb37275642487f1f0
CRs-Fixed: 627278