Changing getPdus to use provided senderAddress to construct PDUs
getPDUs now checks if the provided senderAddress is a well formed SMS-
address. If this is the case, it used in the inital process of creating
the PDUs. If it isn't (for example: it is a vanish- or mail-address),
the MOCK_ADDRESS is initially used and later replaced with the alpha-
numeric string representing the provided address.

JIRA: CYAN-4057
Issue: https://jira.cyanogenmod.org/browse/CYAN-4057
Change-Id: I512ae4074dd538b4394b9e9e5ba479625d2a4393