MSIM(DSDA): Update setMode properly.
When user accepts MT call mSpeedUpAudioForMtCall should be set
to true, but this is not happening when there is already a call active
and user accepts an MT call. As a result when setAudioMode()is called
after user accepting call, audio mode is again set to RINGING.
As part of the change, mSpeedUpAudioForMtCall is set to true,
when user accepts MT call, regardless of whether there is already
an active call or not.

Change-Id: I08cc4955c5d8dbb100bd33b5ef2db2532d43fb56
CRs-Fixed: 690366