Revert "Send SMS over 1x when on eHRPD/LTE and active voice call"
This reverts commit 4eb8e4c8dc684e2109276ea2ec4ce25d3b188568.

The functionality implemented by this change is already there in our internal
change (e87a8e9e7cd6f645a051b5446f27eb547a4de636)
The problem with this change is that it assumes that for a 3gpp2 operator,
the sms format will always be 3gpp2 even on IMS which may not be so.
In such a case, this change won't work. Also, this change has one additional
conditional check for LTE + no concurrent voice and data allowed which is taken
care of by our modem

CRs-Fixed: 750718

Change-Id: Ic953299e5f1db518a3268c5203d51dffbe370408